Crate rusoto_iotanalytics

AWS IoT Analytics allows you to collect large amounts of device data, process messages, and store them. You can then query the data and run sophisticated analytics on it. AWS IoT Analytics enables advanced data exploration through integration with Jupyter Notebooks and data visualization through integration with Amazon QuickSight.

Traditional analytics and business intelligence tools are designed to process structured data. IoT data often comes from devices that record noisy processes (such as temperature, motion, or sound). As a result the data from these devices can have significant gaps, corrupted messages, and false readings that must be cleaned up before analysis can occur. Also, IoT data is often only meaningful in the context of other data from external sources.

AWS IoT Analytics automates the steps required to analyze data from IoT devices. AWS IoT Analytics filters, transforms, and enriches IoT data before storing it in a time-series data store for analysis. You can set up the service to collect only the data you need from your devices, apply mathematical transforms to process the data, and enrich the data with device-specific metadata such as device type and location before storing it. Then, you can analyze your data by running queries using the built-in SQL query engine, or perform more complex analytics and machine learning inference. AWS IoT Analytics includes pre-built models for common IoT use cases so you can answer questions like which devices are about to fail or which customers are at risk of abandoning their wearable devices.

If you're using the service, you're probably looking for IotAnalyticsClient and IotAnalytics.

Structs

AddAttributesActivity
An activity that adds other attributes based on existing attributes in the message.

BatchPutMessageErrorEntry
Contains informations about errors.

BatchPutMessageRequest

BatchPutMessageResponse

CancelPipelineReprocessingRequest

CancelPipelineReprocessingResponse

Channel
A collection of data from an MQTT topic. Channels archive the raw, unprocessed messages before publishing the data to a pipeline.

ChannelActivity
The activity that determines the source of the messages to be processed.

ChannelStatistics
Statistics information about the channel.

ChannelSummary
A summary of information about a channel.

CreateChannelRequest

CreateChannelResponse

CreateDatasetContentRequest

CreateDatasetRequest

CreateDatasetResponse

CreateDatastoreRequest

CreateDatastoreResponse

CreatePipelineRequest

CreatePipelineResponse

Dataset
Information about a data set.

DatasetAction
A "DatasetAction" object specifying the query that creates the data set content.

DatasetContentStatus
The state of the data set and the reason it is in this state.

DatasetEntry
The reference to a data set entry.

DatasetSummary
A summary of information about a data set.

DatasetTrigger
The "DatasetTrigger" that specifies when the data set is automatically updated.

Datastore
Information about a data store.

DatastoreActivity
The 'datastore' activity that specifies where to store the processed data.

DatastoreStatistics
Statistics information about the data store.

DatastoreSummary
A summary of information about a data store.

DeleteChannelRequest

DeleteDatasetContentRequest

DeleteDatasetRequest

DeleteDatastoreRequest

DeletePipelineRequest

DescribeChannelRequest

DescribeChannelResponse

DescribeDatasetRequest

DescribeDatasetResponse

DescribeDatastoreRequest

DescribeDatastoreResponse

DescribeLoggingOptionsRequest

DescribeLoggingOptionsResponse

DescribePipelineRequest

DescribePipelineResponse

DeviceRegistryEnrichActivity
An activity that adds data from the AWS IoT device registry to your message.

DeviceShadowEnrichActivity
An activity that adds information from the AWS IoT Device Shadows service to a message.

EstimatedResourceSize
The estimated size of the resource.

FilterActivity
An activity that filters a message based on its attributes.

GetDatasetContentRequest

GetDatasetContentResponse

IotAnalyticsClient
A client for the AWS IoT Analytics API.

LambdaActivity
An activity that runs a Lambda function to modify the message.

ListChannelsRequest

ListChannelsResponse

ListDatasetsRequest

ListDatasetsResponse

ListDatastoresRequest

ListDatastoresResponse

ListPipelinesRequest

ListPipelinesResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

LoggingOptions
Information about logging options.

MathActivity
An activity that computes an arithmetic expression using the message's attributes.

Message
Information about a message.

Pipeline
Contains information about a pipeline.

PipelineActivity
An activity that performs a transformation on a message.

PipelineSummary
A summary of information about a pipeline.

PutLoggingOptionsRequest

RemoveAttributesActivity
An activity that removes attributes from a message.

ReprocessingSummary
Information about pipeline reprocessing.

RetentionPeriod
How long, in days, message data is kept.

RunPipelineActivityRequest

RunPipelineActivityResponse

SampleChannelDataRequest

SampleChannelDataResponse

Schedule
The schedule for when to trigger an update.

SelectAttributesActivity
Creates a new message using only the specified attributes from the original message.

SqlQueryDatasetAction
The SQL query to modify the message.

StartPipelineReprocessingRequest

StartPipelineReprocessingResponse

Tag
A set of key/value pairs which are used to manage the resource.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdateChannelRequest

UpdateDatasetRequest

UpdateDatastoreRequest

UpdatePipelineRequest

Enums

BatchPutMessageError
Errors returned by BatchPutMessage

CancelPipelineReprocessingError
Errors returned by CancelPipelineReprocessing

CreateChannelError
Errors returned by CreateChannel

CreateDatasetContentError
Errors returned by CreateDatasetContent

CreateDatasetError
Errors returned by CreateDataset

CreateDatastoreError
Errors returned by CreateDatastore

CreatePipelineError
Errors returned by CreatePipeline

DeleteChannelError
Errors returned by DeleteChannel

DeleteDatasetContentError
Errors returned by DeleteDatasetContent

DeleteDatasetError
Errors returned by DeleteDataset

DeleteDatastoreError
Errors returned by DeleteDatastore

DeletePipelineError
Errors returned by DeletePipeline

DescribeChannelError
Errors returned by DescribeChannel

DescribeDatasetError
Errors returned by DescribeDataset

DescribeDatastoreError
Errors returned by DescribeDatastore

DescribeLoggingOptionsError
Errors returned by DescribeLoggingOptions

DescribePipelineError
Errors returned by DescribePipeline

GetDatasetContentError
Errors returned by GetDatasetContent

ListChannelsError
Errors returned by ListChannels

ListDatasetsError
Errors returned by ListDatasets

ListDatastoresError
Errors returned by ListDatastores

ListPipelinesError
Errors returned by ListPipelines

ListTagsForResourceError
Errors returned by ListTagsForResource

PutLoggingOptionsError
Errors returned by PutLoggingOptions

RunPipelineActivityError
Errors returned by RunPipelineActivity

SampleChannelDataError
Errors returned by SampleChannelData

StartPipelineReprocessingError
Errors returned by StartPipelineReprocessing

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateChannelError
Errors returned by UpdateChannel

UpdateDatasetError
Errors returned by UpdateDataset

UpdateDatastoreError
Errors returned by UpdateDatastore

UpdatePipelineError
Errors returned by UpdatePipeline

Traits

IotAnalytics
Trait representing the capabilities of the AWS IoT Analytics API. AWS IoT Analytics clients implement this trait.