Struct rusoto_iotanalytics::IotAnalyticsClient

pub struct IotAnalyticsClient { /* fields omitted */ }

A client for the AWS IoT Analytics API.

Methods

impl IotAnalyticsClient

pub fn new(region: Region) -> IotAnalyticsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> IotAnalyticsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,